+++ examples/node-example.js
@@ -1,6 +1,7 @@
-const SmartLogger = require("../dist/smart-logger");
-const logger = new SmartLogger({ logLevel: "debug" });
+const Logger = require("../dist/logger");
+const logger = new Logger({ logLevel: "trace" });
 
+logger.trace("This is a trace message");
 logger.debug("This is a debug message");
 logger.info("This is an info message");
 logger.warn("This is a warning message");